requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asyncio==3.4.3
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cowsay==6.1
cryptography==42.0.8
cycler==0.12.1
distro==1.9.0
dsnparse==0.2.1
firebase-admin==6.5.0
firecrawl-py==1.2.3
fonttools==4.53.1
git-filter-repo==2.45.0
google-api-core==2.19.0
google-api-python-client==2.133.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
idna==3.7
Jinja2==3.1.4
jiter==0.5.0
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
matrix==3.0.0
msgpack==1.0.8
nest-asyncio==1.6.0
numpy==2.0.1
openai==1.43.0
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
pyparsing==3.1.2
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
setuptools==70.0.0
six==1.16.0
sniffio==1.3.1
termcolor==2.4.0
tqdm==4.66.5
typing_extensions==4.12.2
unimatrix==1.14.0
uritemplate==4.1.1
urllib3==2.2.1
websockets==13.0.1
wheel==0.43.0
